Tilesets folder is back, small graphics code fixes
The user-facing function of the tilesets folder is almost the same - the contents are added to the graphics customisation tab.  On the back end, an installed graphics pack will contain tilesets from vanilla DF (via baselines), the graphics pack, and additional tilesets from the tilesets folder if they exist.  This means less duplication and a more relevant list.  The tilesets folder can therefore be used for any images in /data/art which are shared across packs - it's particularly good for TwbT images.