More fixes!
core.mods.do_merge_seq is going to continue creating PyLint noise - it's a big and fairly complicated function but I can't see a good way to break it up.  I've tried to keep it much more heavily commented to compensate; that and the difflib docs should make it easy enough to follow.