Handle mods reliant on DFHack features, update readme
There are many mods that rely on DFHack, and this upgrades the merge logic to handle them.  Instead of special-case copying a non-vanilla file, we simply use an empty list of lines to stand in for any missing file.  We also correctly handle multiple mods containing the same file (eg a DFHack script), and other common text-file extensions for DFHack (init, lua, rb).

Also update the README, to reflect that the PyLNP is now the definitive DF launcher and does not need to be defined in terms of the old LNP launcher.