Graphics:  distinguish pre- and mid-install aborts, misc fixes
There's an important difference in the return from installing graphics because raws could not be updated (nothing has been touched yet), and returning after an exception when things may be in a bad state.  User feedback now reflects that.

Update raws doesn't need default args, and can return None for invalid pack and False for bad mod combination (though no differences in handling later).  We'll also be careful to clean up if the merge went badly, since we might now use that dir without changes for the next save.  Fixed a typo.