Starting mods merge with graphics now optional.
Also slightly rearranged the buttons at the bottom of the mods tab.

I think this makes the outcome more intuitive in terms of raw compatibility.  There's no indicator on the button, but state is easy to tell from whether there's a graphics pack at the top of the list.  There might be some strange-to-new-users behaviour if a graphics-less mod is installed to DF while using a tileset that relies on non-vanilla raws, but that's a risk in DF generally and already comes up elsewhere as an unavoidable cost of using graphics with mods.