Hide '_' prefixed tilesets; still install them; clarify code.
Corrected docstring and made return statement for read_tilesets() more concise - an implicit generator expression works as well as a list comprehension, with fewer brackets.

I also added a condition to hide files beginning with an underscore, while still copying them to data/art on graphics install.
As documented in the readme, this allows pack maintainers to hide files which are used for TwbT but are not valid tilesets - allowing them to be shared across graphics packs for size savings, while reducing the impact on users.